+++ Assets/Scripts/PlayerNet.cs
@@ -126,7 +126,7 @@ public class PlayerNet : NetworkBehaviour
     [Command]
     public void cmdPlayChallenge(int key, GameObject challenge)
     {
-        BigInteger result = challenge.GetComponent<Challenge>().play(key, gameObject.GetComponent<PlayerManager>().id);
+        int result = challenge.GetComponent<Challenge>().play(key, gameObject.GetComponent<PlayerManager>().id);
         gameObject.GetComponent<PlayerManager>().rcpTargetSetPassword(gameObject.GetComponent<NetworkIdentity>().connectionToClient, result);
 
     }
